[Refactor] [11/N] to simplify the mcp architecture
