+++ vllm/entrypoints/openai/responses/context.py
@@ -22,6 +22,8 @@ from vllm.entrypoints.chat_utils import (
     ChatTemplateContentFormatOption,
 )
 from vllm.entrypoints.constants import MCP_PREFIX
+from vllm.entrypoints.mcp.tool import Tool
+from vllm.entrypoints.mcp.tool_server import ToolServer
 from vllm.entrypoints.openai.engine.protocol import (
     FunctionCall,
 )
@@ -38,9 +40,7 @@ from vllm.entrypoints.openai.responses.protocol import (
     ResponseRawMessageAndToken,
     ResponsesRequest,
 )
-from vllm.entrypoints.responses_utils import construct_tool_dicts
-from vllm.entrypoints.tool import Tool
-from vllm.entrypoints.tool_server import ToolServer
+from vllm.entrypoints.openai.responses.utils import construct_tool_dicts
 from vllm.outputs import RequestOutput
 from vllm.reasoning.abs_reasoning_parsers import ReasoningParser
 from vllm.tokenizers import TokenizerLike
