+++ vllm/entrypoints/openai/responses/utils.py
