+++ vllm/reasoning/gptoss_reasoning_parser.py
@@ -5,12 +5,12 @@ from collections.abc import Sequence
 
 from transformers import PreTrainedTokenizerBase
 
+from vllm.entrypoints.mcp.tool_server import ToolServer
 from vllm.entrypoints.openai.chat_completion.protocol import (
     ChatCompletionRequest,
 )
 from vllm.entrypoints.openai.engine.protocol import DeltaMessage
 from vllm.entrypoints.openai.parser.harmony_utils import parse_chat_output
-from vllm.entrypoints.tool_server import ToolServer
 from vllm.logger import init_logger
 from vllm.reasoning import ReasoningParser
